Added getFormat
Formats are determined by the file extension in the file name.
scrapeMetaData(fileID) function was created as an alternative to determine the mimeType by scraping each google doc page. The function doesn't work with google doc docs as they don't have a mimeType. There are only about 5 google doc docs on the drive so doesn't really matter. Don't see the need to access the meta data using the API.