Added timestamps, versions
Timestamps are now being used. Stored in the "date" field of output.txt. The "year" field is the year if able to be parsed. If able to parse a year from the title then the date aquired from google drive api is overwritten with Jan. 2 of the year parsed as a timestamp.

Google docs currenlty use the current date of running the upload script if unable to parse a date from the title. Waiting for approval of gdoc API

Fixed getVersion in parser

Created requirements.txt and added python-dateutil